backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
databricks-sql-connector==3.0.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
